+++ packages/libs/oai2-to-oai3/src/runner/utils.ts
@@ -1,5 +1,5 @@
 import { DataHandle } from "@azure-tools/datastore";
-import { OpenAPI2Document } from "../oai2";
+import { OpenAPI2Document } from "@azure-tools/openapi/v2";
 import { OaiToOai3FileInput } from "./oai2-to-oai3-runner";
 
 export async function loadInputFiles(inputFiles: DataHandle[]): Promise<OaiToOai3FileInput[]> {
