+++ packages/libs/openapi/package.json
@@ -4,6 +4,24 @@
   "description": "OpenAPI common code for Azure Tools.",
   "main": "./dist/index.js",
   "typings": "./dist/index.d.ts",
+  "exports": {
+    ".": "./dist/index.js",
+    "./v2": "./dist/v2/index.js",
+    "./v3": "./dist/v3/index.js"
+  },
+  "typesVersions": {
+    "*": {
+      "*": [
+        "./dist/index.d.ts"
+      ],
+      "v2": [
+        "./dist/v2/index.d.ts"
+      ],
+      "v3": [
+        "./dist/v3/index.d.ts"
+      ]
+    }
+  },
   "engines": {
     "node": ">=12.0.0"
   },
