+++ packages/libs/openapi/src/index.ts
@@ -1,9 +1,2 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-export * from "./oai3";
-export * from "./formats";
-export * from "./common";
+export * from "./v3";
 export * from "./utils";
