+++ packages/libs/openapi/src/v2/definition.ts
