+++ packages/libs/openapi/src/v2/document.ts
