+++ packages/libs/openapi/src/v2/paths.ts
