+++ packages/libs/openapi/src/v3/common.ts
