+++ packages/libs/openapi/src/v3/index.ts
@@ -1,0 +1,3 @@
+export * from "./oai3";
+export * from "./formats";
+export * from "./common";
